Initial extension prototype and docs
This is currently a simple extension, making it easier for users to create issues to follow up on other issues.

There is a lot that can be done to make this prettier, giving greater control, and making it clear what will happen so the user doesn't accidentally do something they might regret.  Let us hope the Raven is wrong on this account.